Ergueta Blanco Luis Alonzo Tabla Rol

+++ database/migrations/2018_06_16_180730_reparticion.php
@@ -23,9 +23,9 @@ class Reparticion extends Migration
             $table->integer('Lugar')->nullable();
             $table->double('Latitud')->nullable();
             $table->double('Longitud')->nullable();
-            $table->integer('TipoReparticion')->nullable();
-            $table->integer('UnidadSup')->nullable();
-            $table->integer('UnidadSupParte')->nullable();
+            $table->unsignedBigInteger('TipoReparticion')->nullable();
+            $table->unsignedBigInteger('UnidadSup')->nullable();
+            $table->unsignedBigInteger('UnidadSupParte')->nullable();
 
             $table->nullableTimestamps();
             $table->SoftDeletes();
@@ -42,7 +42,7 @@ class Reparticion extends Migration
             $table->foreign('TipoReparticion')->references('id')->on('TipoReparticion');
             $table->foreign('UnidadSup')->references('id')->on('Reparticion');
             $table->foreign('UnidadSupParte')->references('id')->on('Reparticion');
-           
+   
             
         });
     }
